% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{bayescp_validate_fit}
\alias{bayescp_validate_fit}
\title{Validate a fitted ExactBayesCP object}
\usage{
bayescp_validate_fit(object, tolerance = 1e-08, error = FALSE)
}
\arguments{
\item{object}{A `bayescp_fit` object.}

\item{tolerance}{Numerical tolerance.}

\item{error}{If `TRUE`, stop when validation fails.}
}
\value{
A named logical vector.
}
\description{
Validate a fitted ExactBayesCP object
}
