\name{plot.bayescp_fit}\alias{plot.bayescp_fit}\title{Plot a BayesCP Fit}\usage{\method{plot}{bayescp_fit}(x, type = c("posterior", "series", "histogram", "density_mu1", "density_mu2", "density_sigma2", "credible"), ...)}\arguments{\item{x}{A \code{bayescp_fit} object.}\item{type}{Requested plot type.}\item{...}{Additional graphical arguments.}}\description{Produces posterior, series, histogram, parameter-density, and credible-interval plots. Density and histogram plots require posterior draws.}