% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{bayescp_prior_grid}
\alias{bayescp_prior_grid}
\title{Create weak, moderate, and strong prior specifications}
\usage{
bayescp_prior_grid(m01 = 0, m02 = 1, a0 = 2, b0 = 1)
}
\arguments{
\item{m01, m02}{Segment prior means.}

\item{a0, b0}{Inverse-gamma hyperparameters.}
}
\value{
A named list of three `bayescp_prior` objects.
}
\description{
Create weak, moderate, and strong prior specifications
}
\examples{
priors <- bayescp_prior_grid(
  m01 = 0,
  m02 = 1,
  a0 = 2,
  b0 = 1
)

names(priors)
priors$weak
priors$moderate
priors$strong

}
