% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{bayescp_save_plots}
\alias{bayescp_save_plots}
\title{Save standard ExactBayesCP plots}
\usage{
bayescp_save_plots(
  object,
  directory = "bayescp_figures",
  formats = c("png", "pdf"),
  types = c("posterior", "series", "credible"),
  width = 1800,
  height = 1200,
  res = 200
)
}
\arguments{
\item{object}{An object of class `bayescp_fit`.}

\item{directory}{Character string giving the directory in which plot files
should be saved.}

\item{formats}{Character vector containing one or both of `"png"` and
`"pdf"`.}

\item{types}{Character vector specifying the plot types to export.}

\item{width}{Width of PNG output in pixels.}

\item{height}{Height of PNG output in pixels.}

\item{res}{Resolution of PNG output in dots per inch.}
}
\value{
Invisibly returns a character vector containing the paths of created files.
}
\description{
Exports selected graphical summaries from a fitted `bayescp_fit` object
to PNG and/or PDF files.
}
\examples{
dat <- bayescp_simulate(
  n = 100,
  tau = 50,
  mu1 = 0,
  mu2 = 1,
  sigma = 1,
  seed = 123
)

fit <- bayescp_fit(
  dat$y,
  prior = bayescp_prior(
    m01 = 0,
    m02 = 1,
    kappa01 = 5,
    kappa02 = 5
  ),
  draws = 100,
  seed = 123
)

tmp <- tempfile("exactbayescp_plots_")

files <- bayescp_save_plots(
  fit,
  directory = tmp,
  formats = "png",
  types = "posterior",
  width = 800,
  height = 600,
  res = 100
)

files

unlink(tmp, recursive = TRUE)

}
