% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{bayescp_validate_fit}
\alias{bayescp_validate_fit}
\title{Validate a fitted ExactBayesCP object}
\usage{
bayescp_validate_fit(object, tolerance = 1e-08, error = FALSE)
}
\arguments{
\item{object}{A `bayescp_fit` object.}

\item{tolerance}{Numerical tolerance.}

\item{error}{If `TRUE`, stop when validation fails.}
}
\value{
A named logical vector.
}
\description{
Validate a fitted ExactBayesCP object
}
\examples{
dat <- bayescp_simulate(
  n = 100,
  tau = 50,
  mu1 = 0,
  mu2 = 1,
  sigma = 1,
  seed = 123
)

fit <- bayescp_fit(
  dat$y,
  prior = bayescp_prior(
    m01 = 0,
    m02 = 1,
    kappa01 = 5,
    kappa02 = 5
  ),
  draws = 0
)

bayescp_validate_fit(fit)

}
